moved logic out dataset details to component fetch
